+++ amazonorders/__init__.py
@@ -1,3 +1,3 @@
 __copyright__ = "Copyright (c) 2024-2025 Alex Laird"
 __license__ = "MIT"
-__version__ = "4.4.5"
+__version__ = "4.4.6"